% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_miss.R
\name{col_miss}
\alias{col_miss}
\title{col_miss}
\usage{
col_miss(x)
}
\arguments{
\item{x}{A dataframe}
}
\value{
ggplot bar plot, if no missing data returns a message
}
\description{
Plotting missing variables and displaying the proportion
of missing variables by column
}
\examples{
data(mtcars2)
col_miss(mtcars2)
}
