% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{impute: filling in missing data}
\usage{
impute(data, method = c("knn", "missForest", "mice"))
}
\arguments{
\item{data}{data frame with missing values}

\item{method}{method to be applied to impute the missing values}
}
\value{
a new data frame where the missing value are estimated based on the specified method
}
\description{
This \code{impute} function is used to fill in the missing values of a data frame using a specified method
from the following three: knn, missForest or mice. If the method input is not one of the three previously
listed, then the function will throw an error.
}
\details{
The impute function is a wrapper for the following functions:
\href{https://www.rdocumentation.org/packages/class/versions/7.3-20/topics/knn}{knn} function,
\href{https://www.rdocumentation.org/packages/missForest/versions/1.5}{missForest} function,
\href{https://www.rdocumentation.org/packages/mice/versions/3.15.0/topics/mice}{mice} function,
and \href{https://rdrr.io/cran/sjmisc/man/merge_imputations.html}{merge_imputation} function.
}
\examples{
# airquality is a data frame from base R

new_df <- impute(airquality, "knn")
new_df <- impute(airquality, "missForest")
new_df <- impute(airquality, "mice")

# if the specified method is not among the listed three, an error will be reported
new_df <- impute(airquality, "randomforest")
}
