% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_columns.R
\name{plotMiss}
\alias{plotMiss}
\title{plotMiss}
\usage{
plotMiss(x)
}
\arguments{
\item{x}{A dataframe}
}
\value{
ggplot bar plot, if no missing data returns a message
}
\description{
Plotting missing variables and displaying the proportion
of missing variables by column
}
\examples{
\dontrun{
data(sleep, package = "VIM")
plotMiss(sleep)
}
}
