% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_miss.R
\name{row_miss}
\alias{row_miss}
\title{Row Chart}
\usage{
row_miss(df)
}
\arguments{
\item{df}{}
}
\value{
bar chart of missing values per case
}
\description{
This \code{row_miss} returns a bar chart that shows you the missing values per case.
}
\examples{
data(mtcars2)
row_miss(mtcars2)
}
