% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_miss.R
\name{col_miss}
\alias{col_miss}
\title{Plot missing data by variable}
\usage{
col_miss(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
ggplot2 graph
}
\description{
Plot the amount of missing data by variable
}
\examples{
data(mtcars2)
col_miss(mtcars2)
}
