% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_miss.R
\name{row_miss}
\alias{row_miss}
\title{Plot missing data by row}
\usage{
row_miss(data)
}
\arguments{
\item{data}{}
}
\value{
ggplot2 graph
}
\description{
Plotting the amount of missing data by observation
}
\examples{
data(mtcars2)
row_miss(mtcars2)
}
